UPD: More changes to anchors and spacing in Options.
Each top control (group boxes) in a frame must be anchored to parent or not anchored at all, they cannot be anchored to one another because autosizing in the scrollbox won't work correctly. Each control in an autosized container must be anchored so that the container is autosized properly. Using VerticalSpacing on TFrame currently does not work correctly with autosize, so replaced with spacing of individual controls.

git-svn-id: https://svn.code.sf.net/p/doublecmd/code@3989 10bb5c37-c6d6-4c74-9a2a-30fc9b6e22f7